First pass at functioning CLI
Why these changes are being introduced:

This builds on the project scaffolding to provide a CLI capable
of preparing the notebook environment and running the notebook.

This does NOT yet contain much of any meaningful testing.

How this addresses that need:
* Establishes CLI command and signature
* Supports a pre-existing ("mounted") notebook location or
a git clone for a repository
* Prepares a full notebook path to run
* Prepares a marimo subprocess command that handles
inlined or external dependencies
* Whether inline or external dependencies, the process
is always isolated via `uv` given the `--sandbox` or
`--with-requirements` flags.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1415