Use explicit auth token or none
Why these changes are being introduced:

With the --no-token or --token flag, it was a bit confusing
and it did not allow for setting an explicit token for the noteobok.
We need to support setting the token such that we can know it
advance.

How this addresses that need:
* Changes the CLI arg to `--token:str|None` which when None
sets `--no-token` in marimo launch else sets the explicit token
passed

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1415