Support launching the notebook with a base URL
Why these changes are being introduced:

Based on the interplay of the ALB and the notebook running in an ECS
fargate container, we realized that we would need to launch the notebook on
a particular URL path.

Thankfully, Marimo supports this with a --base-url flag.

How this addresses that need:

* Adds --base-url / NOTEBOOK_BASE_URL configurations to explicitly set a
base url to launch on
* Adds flags + env vars to skip a base URL entirely
* If neither are passed, the default behavior is to construct a base URL
from the repository name + notebook path

Side effects of this change:
* This aligns with the URL patterns that ALB requests will send to the running
notebook.  By passing --repo https://github.com/foo/bar and --path super/notebook.py,
the resulting base URL will be "bar/super/notebook.py" which is precisely
what the ALB will send as requests

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1441