Explicit base URL or None
Why these changes are being introduced:

We realized we could simplify this a bit at the infrastructure level
and *always* pass a base URL if needed.  This means we can simplify
this application to only apply a base URL when explicitly passed.

How this addresses that need:
* Removes any notion of "skipping" a base URL
* CLI arg / env var passes along an explicit base URL, or none is set

Side effects of this change:
* ALB + ECS service infra is responsible for knowing and setting the
base URL

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1441