Update Workflows
Why these changes are being introduced:
This is the first step in migrating this repository to optionally build
for either AMD64 or ARM64 CPU Architecture for deployment in AWS.

How this addresses that need:
* Update the `Makefile` with the new output from the
mitlib-tf-workloads-ecr repository (and then make some further
modifications that will be reflected back in the mitlib-tf-workloads-ecr
repository soon)
* Update the dev-build.yml workflow with the new output from the
mitlib-tf-workloads-ecr repository
* Update the stage-build.yml workflow with the new output from the
mitlib-tf-workloads-ecr repository
* Update the prod-promote.yml workflow with the new output from the
mitlib-tf-workloads-ecr repository
* Update the README with notes about building and deploying in AWS

Side effects of this change:
None. Since we did not create an `.aws-architecture` file, the builds
for AWS will still default to AMD64 as before.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1448